refactor(ocr): extract engine selection factory into engine/factory
